clk: keystone - add Main PLL clock driver
This is the driver for the main PLL clock hardware found on Keystone
devices (c6x such as C6678). This driver is implemented as per common
clock provider API. The main PLL hardware typically has a multiplier,
and a divider. struct clk_keystone_pll_data is used to configure the
driver for a specific platform.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>