ARM: davinci - adding new type for cpu_clks in soc_info
To define the clock nodes available on a SoC, a new structure is
defined that will be used by SoC specific code to define the clock
table. This table is used by davinci_common_clk_init() to initialize
the clk drivers. Also added a new field dev_clk_lookups to define
clock aliases for devices that share same clocks.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>