hwqueue: extend push and pop api's to take an additional argument
An additional argument is being added to the hardware queue push and
pop api's.
This is needed due to the fact that there are implementations like
that of keystone hardware queue which need both a descriptor size
as well as packet size in a push operation.
A new argument called "flags" is now an additional argument to the
hardware queue push api. This can be used in a way defined by any driver
that leverages the hardware queue push api.The upper bits can have some
encoded information and the lower bits can have actual data.
The corresponding pop api too has this same additional
argument though this additional argument has no use at present.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>