dma: keystone: use extended push and pop hwqueue api's
The hardware queue push and pop api's wer extended to take an additional
argument.This is now being leveraged in the keystone dma engine.
An additional argument called flags is passed into the hwqueue push api.
If we want packet size to be sent to the lower layer, we extract the
appropriate bits from the descriptor info and encode it with an
appropriate flag that can be understood by the keystone hardware
queue layer. If we don't want the packet size to be sent to the lower
laters, we simply pass a 0. The pop api's took take and
additional argument. We pass a 0 for the flags.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>