net: keystone: Improve TX flow control handling
Prior to this patch, Linux interface queues were only flow-
controlled after a failure to map an SKB for transmission.
This meant the packet was dropped. A potential near-deadlock
condition could arise because descriptors were only reclaimed
after a transmit occurred, and if all interface queues were
XOFFed then the only mechanism for recovery was the TX timeout.

This patch introduces a proactive flow control mechanism.
Rather than tracking packet counts, we now track descriptor
usage. If the number of free descriptors drops below the
pause threshold (which should be high enough to avoid dropping
the triggering packet), the interface queue is XOFFed and
we switch to reclaiming descriptors in a tasklet scheduled from
the dmaengine notify routine (interrupt). When the number of
free descriptors rises above the resume threshold, the interface
queue is restarted.

This mechanism works best if each interface queue is mapped 1:1
with a txpipe. This avoids having to XOFF multiple interface
queues separately. However, it should still work if this is not
the case, with an increased chance of dropping packets that
could not be mapped.