net: keystone: Fix ndo_do_ioctl method return status
If we are called with an ioctl code we don't recognize, we need
to reject it with -EOPNOTSUPP. Instead we were returning 0, which
basically meant all ioctls succeeded even if we didn't support it.
This caused the LLDP daemon to think we were a bonding interface.

Note that this has the potential to cause applications to break,
since we don't support the normal MII ioctls. However, adding
support for MII ioctls would be a significant effort.

Signed-off-by: Reece Pollack <x0183204@ti.com>