dma: keystone: allow flow tag override at prep
This patch allows the dma-engine user layer to override the flowtag on a per
packet basis at transfer prep time.  The expectation is that the caller passes
in an "options" argument with the DMA_HAS_FLOWTAG flag set, and with the
flowtag information included (see DMA_FLOWTAG_MASK and DMA_FLOWTAG_SHIFT).

Signed-off-by: Cyril Chemparathy <cyril@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>