net: keystone: fix dma addressing in scatter list
On LPAE systems, the DMA address is no longer necessarily the physical
address.  On Keystone platforms, the physical address belongs to the
08:0000:0000 range, which then gets mapped down to the 00:8000:0000 range by
the SES MPAX.

The following patch fixes an odd instance where the code was incorrectly
filling up a physical address into a scatterlist before doing a DMA unmap
operation.  This causes invalid page access crashes when running the network
driver with LPAE.

Signed-off-by: Cyril Chemparathy <cyril@ti.com>