tci6638: dts: add device tree bindings for multi qos trees
this commit adds the necessary device tree bindings to
demonstrate qos support in each interface. Added in this
patch are an example qos tree for interface-1 which is
the same as that used for interface-0. Also, we need
qos channels for the linux qos module. These bindings
are also added as part of this patch.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>