net: keystone: disable VLAN-awareness in the 1G and 10G switches
The current VLAN-aware configuration does bad things to incoming
packets when in multi-interface mode. On the tci6614, packets
containing VLAN tags have the tags stripped off. On the tci6638,
packets that don't have VLAN tags have tags added. Neither one
is appropriate.

Since there is no reason for the switch to be VLAN-aware, this
patch disables VLAN aware mode. It does NOT, however, remove
the ALE entries for the VLANs. More research should be done to
determine the proper configuration when time permits.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>