net: ethernet: cpts: reduce overflow period
If an ethernet event occurs around a counter rollover, the rollover
event could possibly be loaded into the event FIFO before the ethernet
event is loaded into the event FIFO even though the ethernet event time
was actually taken before the rollover.  This misaligned event condition
arises because an ethenet evnet time stamp ocurs at the beginning of a
packet and time passes before the packet is determined to be a valid
sync packet. The misaligned evnet condition occurs if the rollover occurs
in the middle, after the packet time stamp has been taken, but before the
packet has been determined to be a valid time sync packet.

Reduced overflow period essentially reduces the timecounter refresh rate
and prevents the occuring of misalignment events.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>