net: keystone: cpts: fix PTP message offset calculation
The offset of a PTP message in an IPv4 message over ethernet frame is
the sum of the  ethernet header size, the IP header size and the UDP header size.
When vlan-tag exits, the IPV4 header length field is shifted down by the
number of bytes of the vlan-tag size.  This commit takes care of this shift in
such case.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>