tci6638: dts: enable dma coherency for hwqueue
In the kernel Keystone architecture implementation, dma cache coherency
is a device specific attribute. If the DMA for the device is cache
coherent with the CPU, then the "dma-coherent" attribute has to be
defined in the device tree for specific device node. Based on this
attribute the dma operations for that device are mapped appropriately to
coherent or non-coherent dma operations. This patch enables DMA cache
coherency for hwqueue driver.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>