net: keystone: Allocate memory in atomic contexts using GFP_ATOMIC
This patch replaces one instance where memory is being allocated in
an atomic context to use GFP_ATOMIC rather than GFP_KERNEL, as the
latter may attempt to sleep.

A second allocation is removed entirely and replaced with auto-class
variables, as it's only two u32 vars that are only used locally. In
this case, also added a check for bogus interface types.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>