net: keystone: dma-coherency fix
In the kernel Keystone architecture implementation, dma cache coherency is a
device specific attribute. If the DMA for the device is cache coherent with the
CPU, then the "dma-coherent" attribute has to be defined in the device tree for
specific device node. Based on this attribute the dma operations for that
device are mapped appropriately to coherent or non-coherent dma operations.
For this scheme to work, the device driver needs to pass the correct
"struct device" parameter to the dma functions.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>