pci: keystone: move platform setup code to driver folder
serdes initialization and RC mode setting are specific to a
platform. This patch adds a K2 specific platform setup code
for PCIE and move the same to the driver folder. To support
PCI on another platform/device user may add a new match id string
and implement another platform setup code function for the
new device through this mechanism. If platform code is same,
the bindings for PCI can be same as in K2.

This patch also enable PCI link training based on DTS
attribute "enable-linktrain" to disable link training when
boot loader does the same.

Also add a Documentation for the PCIE bindings.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>