net: keystone: Fix Tx HW checksum offload
This patch fixes a bug in the NETCP PA driver for Tx HW checksum offload.
If the size of TCP/UDP data to be checksummed is odd, a zero octet needs to be
padded to the checksum calculation. NETCP HW supports checksumming odd data
size, software doesn't need to do the zero padding for this. This patch removes
the code in keystone NETCP PA driver which was adding a pad byte for packets
where checksumming has to be done on odd size. Moreever the code had a bug that
it was not updating the skb->len after adding the pad byte.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>