[tmp] clocksource: arch_timer: Remove C3_STOP to make tickless work
On keystone SOC, we don't support CPUIdle power managements and hence
doesn't need the timers to be marked with C3_STOP. With C3_STOP, the
tickless kernel breaks breaking the kernel timer resolution in absence of
broadcast source.

Lets remove the C3_STOP for now to fix the resolution issue.

Revisit it when CPUIdle functionality comes into equation.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>