hwqueue: keystone: Support negative QoS overhead-bytes attribute
This patch supports negative QoS overhead-bytes attributes in the
device tree and via sysfs. This is intended to allow subtracting
a packet header size (ex: -20 for the IP header) from the output
data rate.

To specify a negative value in the device tree it may be necessary
to wrap the value with paretheses, as in this example:
	overhead-bytes = <(-20)>;
The paretheses are not required or accepted when writing to the
sysfs attribute. Also, changing a parent's sysfs attribute does not
cause the new value to propagate to its children; each child must
be changed separately.

This patch requires a new version of the QoS firmware (2_0_1_6), which
in turn requires changes to the firmware name in the device tree.
Without these changes, setting "overhead-bytes" negative behaves as
if it was set to zero. No error is reported.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>