rapidio: fix inconsistency in channel number type
This patch ensures that channel number (AKA channel ID) is specified as u16
in all data structure and functions calls that use it. This matches to
source and destination channel ID fields in channel message header and
is aligned with user space PDRIO library.

Because of switching to strict use of u16 type for channel ID, reserved
value 0 is used to specify dynamic channel allocation request (instead of -1).
As a result of this change all depending upper level code has to be
adjusted accordingly.

Signed-off-by: Alexandre Bounine <alexandre.bounine@idt.com>