hwqueue: keystone: Make writable QoS attributes truly dynamic
Prior to this patch, writing any qos-tree attribute through sysfs caused
the entire QoS configuration to be cleared and restarted. This could cause
undesirable effects on queued traffic. This patch makes all writable
QoS parameters changable without restarting the configuration.

Also corrected in this patch are bugs in how the output-rate parameter
is transformed into sched_cir_credit and sched_out_throttle. The value
of output-rate for a node (port) should affect the sched_cir_credit
of the given node, and the sched_out_throttle of all its children.
Prior to this patch, it set the sched_out_throttle of the given node.

Finally, write permissions for the input-queues parameter have been
removed. Changing this also changes the structure of the QoS tree,
which is not intended to be supported. The "store" method remains
in the code in case this functionality needs to be restored, but it
should not be accessible without also changing the permissions.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>