tci6638: dts: Change eth1 QoS output queue to 645
The output of the QoS tree for eth1 (qos-tree-2) was going to queue 648,
which is the queue for the CPSW. This bypasses PA PDSP5, which performs
essential functions such as checksum offload and timestamping. With
checksum offload enabled, this caused outgoing packets to have invalid
TCP and UDP checksums.

This patch changes the qos-tree-2 output-queue attribute from 648 to 645,
which is the queue for PA PDSP5.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>