tci6638: dts: Configure QoS Bypass as subqueue 0 (the default)
This patch reconfigures the NetCP interface subqueues so that the default
subqueue (0) submits directly to the PA (queue 645, "qos-bypass-X"). This
bypasses QoS in the absence of a TC configuration. If QoS is desired, TC
can be used to feed packets into the other six subqueues (1-6).

The number of descriptors allocated to the QoSo bypass subqueue is also
increased from 64 to 128, as this appears minimally sufficient to achieve
line rate (1 Gbps) transmission.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>