net: keystone: Implement "mark_mcast_match" special packet processing
This patch implements special packet processing for specific marked packets.
The intended use is:
  1) Configured in multiple-interface mode
  2) CPSW ALE re-enabled via /sys/class/net/eth0/device/ale_control
  3) NetCP interfaces slaved to a bridge
  4) NetCP interfaces feed a common QoS tree
  5) Bridge forwarding disabled via "ebtables -P FORWARD DROP"

In this rather odd situation, the bridge will transmit multicast (and
broadcast) packets by sending one on each of the slaved interfaces. This
results in multiple packets through QoS, which is considered "bad".
To avoid this, multicast packets can be marked via ebtables for special
processing in the NetCP PA module before the packets are queued for
transmission. Packets thus recognized are NOT marked for egress via a
specific slave port, and thus will be transmitted through all slave ports
by the ALE.

To do this, a new DTS parameter "mark_mcast_match" has been added. This
parameter takes two u32 values: a "match" value and a "mask" value.

When the NetCP PA module encounters a packet with a non-zero skb->mark
field, it bitwise-ANDs the skb->mark value with the "mask" value and
then compares the result with the "match" value. If these do not match,
the mark is ignored and the packet is processed normally.

However, if the "match" value matches, then the low-order 8 bits of the
skb->mark field is used as a bitmask to determine whether the packet
should be dropped. If the packet would normally have been directed to
slave port 1, then bit 0 of skb->mark is checked; slave port 2 checks
bit 1, etc. If the bit is set, then the packet is enqueued for ALE
processing; if the bit is NOT set, the packet is silently dropped.

An example...

The device tree contains this PA definition:

	mark_mcast_match	= <0x12345a00 0xffffff00>;

The runtime configuration scripts execute this command:

	ebtables -A OUTPUT -d Multicast -j mark \
		--mark-set 0x12345a01 --mark-target ACCEPT

When the bridge attempts to send an ARP (broadcast) packet, it will send
one packet to each of the slave interfaces. The packet sent by the bridge
to slave interface eth0 (CPSW slave port 1) will be passed to the CPSW,
and the ALE will broadcast this packet on all slave ports. The packets
sent by the bridge to other slave interfaces (eth1, CPSW slave port 2)
will be silently dropped.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>