net: keystone: cpts: clean up event queue only when it is full
When events are read from cpts's fifo, it is queued in a driver
event queue with a timeout value.  PPS (TS_COMP) event is
occasionally deleted from driver's event queue before it is
reported to user application because the event queue is checked
for expired event each time a ptp message is received or transmitted.

It is modified in this patch that expired events are only removed
when the driver event queue is full and more free event entries are
needed for newly received events.

Also TS_COMP events are allocated with a longer timeout, i.e., twice
that of the overflow check period.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>