rapidio: rework device hierarchy and introduce mport class of devices
This patch removes an artificial RapidIO bus root device and establishes
the actual device hierarchy. It also introduces device class for RapidIO
mports.

The above changes will simplify handling of device insertion and removal
and notification of device drivers.

This is preliminary version to be used as a proof of concept.
Additional changes are planned for the final release version.
TODO:
- Review use of (need for) internal mport and device lists.
  Consider replacement of them by bus, class and device iterators.
- Review locking.

Signed-off-by: Alexandre Bounine <alexandre.bounine@idt.com>