net: keystone: Remove synchronization between xmit submit and complete
Prior to this patch, it was necessary to block execution of the TX
competion processing during the call to dmaengine_submit(). This was
because the completion function needed the "dma cookie" which was
returned by the submit call, but the packet could be submitted,
processed, and delivered to the completion routine before the submit
call returned.

By recognizing that the TX completion function doesn't need to check
the DMA engine status, the need to use the cookie in the completion
function goes away. This allows the locking around the submit call
to go away too. A significant performance increase results.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>