net: keystone: Don't stop/wake TX queues in timer function
This patch removes the manipulation of TX queue flow control from
the cpsw_timer() function.

Prior to this patch, if the cpsw_timer() function detected that
TX subqueue 0 was stopped while the link is up, it arbitrarily
restarted the subqueue. Since the queue is most likely stopped
because we're near to exhausting the descriptor pool, this is
the wrong thing to do and may result in a mapping failure.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>