aer: add command line option to disable MSI for AER driver
keystone pcie sub system uses platform irq for aer error. So
this requires a way to disable msi irq. the dev->irq traditionally
is initialized to legacy irq for this root port. The keystone
pcie driver adds a fix up to overwrite this irq with a platform
irq so that error handler is attached to it

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>