net: keystone: Allocate pinfo structs from kmem cache in TX path
Every packet passing through the NetCP core is associated with a
netcp_packet (or "pinfo") structure. Long ago, the RX path started
using a kmem_cache to manage allocation of these structures for
performance reasons. This patch changes the TX path to use the same
kmem_cache for its allocations. This results in a 3.6% increase in
measured throughput.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>