net: keystone: Reorder RX path init to fix startup race
The RX notification routine was being registered before the NAPI
instance was enabled. Unlike a tasklet, an attempt to schedule a
disabled NAPI instance is ignored.

This patch reorders the initialization code so this can't happen.
It also moves the netcp_rx_notify() function closer to the rest
of the RX path code to make it easier to find.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>