net: keystone: Make psdata functions static inline
The psdata functions are short helpers to manipulate the PA command
stack stored in the psdata array of the netcp_packet structure. These
are often called with compile-time constants as sizes. By making them
static inline instead of exported, much of the code is removed by the
compiler optimizer, saving precious cycles for every packet transmitted.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>