net: keystone: cpts: add support of ignoring frequency adjust
This patch adds support of bool config cpts-ignore-adjfreq from
devicetress.  When cpts-ignore-adjfreq is set, cpts will skip
adjusting the ptp clock frequency when the ptp adjfreq API is
called.  This is mainly for testing purpose because some cpts
ref clock's frequency can't be adjusted and test program still
calls the adjfreq API which affects the accuracy of the reported
pps timestamping time.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>