net: keystone: Optimize the PA TX Hook
The biggest change is to restructure pa_fmtcmd_next_route() to avoid
constructing a pa_cmd_next_route structure which is used only to
format the actual PA next route command. This function now directly
formats next route command from the netcp_packet structure and the
ps_flags parameter.

All of the pa_fmtcmd_* functions are now marked "inline" for best
performance.

The macros used to format PA commands expand to a series of bitwise
AND and OR operations. By explicitly setting each structure element
to zero within view of the compiler before invoking these macros, the
optimizer collapses many of these separate AND/OR operations into a
single store of a constant.

NOTE that this change requires all users of netcp_push_psdata() to
initialize the allocated area, as this function no longer zeros it.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>