net: keystone: xge: reset serdes before configuration
If xge serdes is in operation before configuration, such as
after a soft reset of the SOC, reset the serdes first.  Otherwise
it may cause a latch-up issue or damage internal power supplies.

Also added is a time out period when checking whether the serdes
pll is locked or not.  When serdes is not locked, XGE driver should
not continue to other configurations.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>