net: keystone: Build SKB in netcp_rx_complete vs netcp_rxpool_alloc
This defers the allocation and initialization of an sk_buff structure
to netcp_rx_complete() instead of netcp_rxpool_alloc(). This gives a
significant performance benefit by avoiding allocating cache lines
with the sk_buff content which likely get flushed and reloaded later.
By using build_skb() to allocate and initialize the sk_buff just before
pushing the packet into the network stack, the cache lines are hot
for their use in RX processing.

The skbuff code maintains a cache of network buffers. Allocating from
this cache using netdev_alloc_frag() is much more efficient than going
to the slab allocator kmalloc(), but is limited to page sized (or less)
allocations. This also requires us to be smart about freeing the buffer
later so that it is released using the correct function.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>