dma: keystone: Optimize copying and scatterlist updates
This patch changes the chan_prep() and chan_complete() functions
to minimize the copying of descriptor fields. Previously, the
PSINFO and EPIB buffers, if present, were copied to the descriptor
in chan_prep(), and copied back in chan_complete(). Now they're
copied in chan_prep() only for MEM-to-DEV channels, and in
chan_complete() only for DEV-to-MEM channels. Also, chan_complete()
touches the scatterlist only for DEV-to-MEM channels, and never
for MEM-to-DEV channels.

The effect of these changes is to allow DEV-to-MEM channels to
use non-persistent scatterlists and EPIB and PSINFO buffers, as
these are never referenced after chan_prep_slave_sg() returns.
This simplifies the design of the completion routine.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>