net: keystone: Avoid allocating a netcp_packet for TX packets
With the companion patch to keystone-pktdma in place, we can avoid
the overhead of allocating and freeing a netcp_packet structure
for each transmitted packet. The TX hooks still require one, but
this can be allocated on the stack in the ndo_start_xmit method.
This should reduce the per-packet overhead, and the SKB handling
is a step toward eliminating the dmaengine layer from this driver.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>