net: keystone: Fix BUG when last fragment is < 4 bytes long
The buffers received from the NetCP includes the 4-byte FCS at the
end of the packet. This makes some upstream code unhappy, so we need
to remove it in the driver. The previous patch for this assumed that
the last fragment would be at least 4 bytes long. This is not a safe
assumption, and when it's incorrect the system crashes in __skb_pull().

This patch fixes this problem by fully assembling the packet first,
then adjusting the packet length using __pskb_trim().

Signed-off-by: Reece R. Pollack <x0183204@ti.com>