rapidio: add support for static enumeration and early registration of hostd id
Static enumeration allows to discover the RapidIO topology without waiting
for beeing enumerated by a remote host and using the remote host id
instead of dynamically creating one like with standard enumeration.

Each node should specify its own distinct riohdid greater or equal to
ero in the kernel command line and set the 'static_enum' rio-scan
module parameter to 1 (default is off).
The enumeration will then be performed by reading the host id and the
component tag of each node. This is similar to the discovery mechanism
but without the need of being enumerated by a peer node.

For this purpose the host id must be set in the CSR registers at
registration time, not when starting the enumeration/discovery.

It is recommended not to use this feature with automatic enumeration
at boot time (scan = 1) but with scanning initiated from user space
in order to wait that all nodes have set their host id.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>