hwqueue: keystone: add qpend queue module param in hwqueue test
In the hwqueue_test module, a qpend queue is opened for push/pop
test.  But the qpend queue number is hard coded and may not be
applicable on every keystone platform.  This patch adds a module
parameter "qpend=N" to specify the qpend queue N to be used in
the test.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>