net: keystone: sa: ipsec inflow issues on egress path with sa context stored in skb.
If ipsec inflow is enabled, for egress packets the sa context structure
is allocated and carried in the skb. The sa context structure needs to
be updated to include a reference count so that it wont be double freed
in case the skb is cloned, as well as a magic number/id field in order
for the sa_tx_hook to differeniate the sa context from a kernel structure
kernel structure that is used whenever ipsec is used on the receive path.

Signed-off-by: Tinku Mannan <tmannan@ti.com>
Signed-off-by: Hao Zhang <hzhang@ti.com>