net: keystone: ale: explicitly disable ALE forwarding on all ports at start up
During the start up of the ethernet switch driver, the switch is not reset and
unused ports are not reconfigured.  Hence the forwarding state of these unused
ports are inherited from u-boot's configuration which is undesirable.  This
patch adds code to disable the forwarding state of all ports during the start
up of the ALE driver.  The forwarding state is enabled when a port is
configured.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>