hwqueue: keystone: Support joint-lite QoS port pairs
QoS firmware 2.0.1.7 supports the concept of "joint" ports, where
an even/odd pair of lite ports work as if they are a single double-
wide port, managing 8 instead of 4 input queues.

This patch adds configuration support for this to the Hardware
Queue QoS module.

No new configuration keywords are necessary to take advantage of
this feature. Configuring a WRR or SP port with more than 4 inputs
will automatically result in the allocation of a joint port pair.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>