hwqueue: keystone: Bounds-check the output-rate burst size
If the second parameter of the output-rate parameter (or the sysfs
burst_size attribute) is set too large, the firmware's calculation
of the circuit credit will roll negative. This causes unwanted
packet drops.

This patch applies bounds-checking to these parameters. If the
device tree contains a bad value we cap it at the calculated
maximum value and output a warning. If a bad value is written to
sysfs, we reject it by returning an error.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>