hwqueue: keystone: Support "blended-scheduler" QoS node type
This patch adds support for the "blended-scheduler" QoS node type,
which allows a mix of "strict-priority" and "weighted-round-robin"
functionality in a single port. Child nodes containing a "priority"
attribute are treated as SP inputs, while child nodes with a
"weight" attribute are treated as WRR inputs. SP inputs are given
priority above WRR inputs.

Children with a new attribute, "low-priority", are treated as SP
inputs, except they are processed only when all SP and WRR input
queues are empty.

This exposes the full functionality of the underlying firmware and
can permit significant reduction in port usage.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>