keystone: dts: move QoS specific nodes to EVM dts file
Network QoS related parameters doesn't belong to the common SoC
specific DTS file. This is an application and should be defined
at a board level. A specific implementation may not use QoS at all or
use it on a specific interface based on use case. So move this to
to board specific dts file.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>