gpio: keystone: update for second instance of gpio
On Lamarr, there are two instances of GPIO. Current code assumes
that address space 0x00 - 0x10 that has binten register is common
across all instances of GPIO. But this is not true and every GPIO
has its own memory map for this. So this is fixed to have common
register space.

Also changed __raw_readl/writel() APIs to readl/writel(). Update the
devicetree bindings documentation to reflect the reg property value
change.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>