dma: keystone: Clean up close handling, add dma_rxfree_flush()
This patch cleans up the handling of channel close when packets
may be still in-flight. It also resolves a possible deadlock
if chan_destroy() is called while the channel is in OPENING
state.

It also adds a new control function, dma_rxfree_flush(), to allow
higher level drivers to request the flushing of the RX Free pool
on a channel. While this is still done automatically during channel
close, there are occasions when it's desirable to do this BEFORE
closing the channel.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>