net: keystone: Clean up network interface shutdown
This patch cleans up netcp_ndo_stop() to better handle interface
shutdown while there is network traffic flowing. It reorders the
shutdown sequence to flush the RX Free pools before closing the
DMA RX channel to stop the flow of incoming packets.

This patch also asserts SGMII RTRESET during shutdown to avoid
having the hardware wedge when shutting down with high incoming
traffic rates. This is cleared when the interface is brought
back up.

An implementation flaw in the management of PA v1 PDSP1 LUT1 table
entries caused crashes with repeated bouncing of interfaces. This
patch corrects that as well.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>