hwqueue: keystone: Fix a bug in bounds-checking sysfs writes
In both qnode_output_rate_store() and qnode_burst_size_store() it's
possible for an invalid value to be detected after it's been stored
into the qnode field. This does not result in the PDSP being updated,
but it makes the display of the attributes wrong. This patch fixes
this.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>