ARM: keystone: remove hardcoding of number of possible CPUs
Currently on K2L boot up log shows

[  641.824274] CPU2: failed to boot: -2
[  641.824406] CPU3: failed to boot: -2

This is because keystone_smp_init_cpus() calls set_cpu_possible()
for N_CPUS which is set to 4, the maximum possible CPUs. nr_cpu_ids
gets initialized with number of available CPUs based on DT nodes and
there is no need for implementing smp_init_cpus(). Also remove
smp_prepare_cpus init code as it is not doing anything. This is in
sync with latest upstream code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>